# dicesGame/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Dice Game - OpenAPI 3.0",
        "description": "API Rest about a game that consists of 2 dice. If the result is 7, you win the game; otherwise, you lose it.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "gmaureirapalma@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/players/{id}/games": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Games list for a specific player",
                "description": "This endpoint is used to retrieve the list of games for a specific player.",
                "operationId": "72a7f2a04946ba396e94ac7858584fb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "player": {
                                            "description": "Player name",
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "games": {
                                            "description": "List of games",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Game"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "Player not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "A specific player throw the dice",
                "description": "This endpoint is used to simulate a player throwing the dice.",
                "operationId": "5d8d8c47d882e99a76ea76d37179f69e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "description": "Result of the dice throw (win/lose)",
                                            "type": "string",
                                            "example": "win"
                                        },
                                        "dice_1": {
                                            "description": "Value of the first dice",
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "dice_2": {
                                            "description": "Value of the second dice",
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "Player not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "A specific player delete all the games",
                "description": "This endpoint is used to delete all the games of a specific player.",
                "operationId": "6f71db3701c13acb3c553c6adaf658c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "Games have been deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "Player not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/players/ranking": {
            "get": {
                "tags": [
                    "Players, Administrator"
                ],
                "summary": "Player's list with better puntuation first",
                "description": "This endpoint is used to retrieve the list of players, ordered by better puntuation first.",
                "operationId": "abadf4bf0b9392da1e399b6123dc5128",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "players_list": {
                                            "description": "List of players",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "Player's name",
                                                        "type": "string"
                                                    },
                                                    "success_rate": {
                                                        "description": "Player's success rate",
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/players/ranking/winner": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Get player with highest success rate",
                "description": "This endpoint retrieves the player with the highest success rate.",
                "operationId": "226979a9b0210769e0bece16814d77ec",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "player": {
                                            "description": "Name of the player",
                                            "type": "string"
                                        },
                                        "success_rate": {
                                            "description": "Success rate of the player",
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/players/ranking/loser": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "Get player with worst success rate",
                "description": "This endpoint retrieves the player with the worst success rate.",
                "operationId": "ecdeced4008d85ff441a7c6bcdd10432",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "player": {
                                            "description": "Name of the player",
                                            "type": "string"
                                        },
                                        "success_rate": {
                                            "description": "Success rate of the player",
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Game": {
                "title": "Game",
                "description": "Game model",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "dice_1": {
                        "description": "Value of dice 1",
                        "type": "integer",
                        "example": 4
                    },
                    "dice_2": {
                        "description": "Value of dice 2",
                        "type": "integer",
                        "example": 2
                    },
                    "is_win": {
                        "description": "Indicates if the game is a win or not",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-19T12:00:00+00:00"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-19T13:30:00+00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer token here)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}